requirements.txt
fastapi
uvicorn
lark-oapi
jinja2
weasyprint
python-dotenv